+++ vscode/src/release.ts
@@ -1,4 +1,5 @@
 import { CodyIDE } from '@sourcegraph/cody-shared'
+import { SG_BLOG_URL } from './chat/protocol'
 
 type ReleaseType = 'stable' | 'insiders'
 
@@ -45,9 +46,7 @@ export function getReleaseNotesURLByIDE(version: string, IDE: CodyIDE): string {
 
     switch (IDE) {
         case CodyIDE.VSCode:
-            return isStable
-                ? `https://github.com/sourcegraph/cody/releases/tag/vscode-v${version}`
-                : 'https://github.com/sourcegraph/cody/blob/main/vscode/CHANGELOG.md'
+            return getReleaseBlogPostURL(version, IDE)
 
         case CodyIDE.JetBrains:
             return isStable
@@ -55,6 +54,32 @@ export function getReleaseNotesURLByIDE(version: string, IDE: CodyIDE): string {
                 : 'https://github.com/sourcegraph/jetbrains/releases'
 
         default:
-            throw new Error('IDE not supported')
+            throw new Error(`No release note for ${IDE}.`)
     }
+}
+
+/**
+ * Gets the release blog post URL for the given IDE and version.
+ *
+ * This function constructs the release blog post URL for the given IDE and version.
+ * For VS Code, it constructs the blog URL based on the version number.
+ *
+ * @param version - The version of the IDE.
+ * @param IDE - The IDE to get the release blog post URL for.
+ * @returns The release blog post URL for the given IDE and version.
+ */
+function getReleaseBlogPostURL(version: string, IDE: CodyIDE): string {
+    const blogURL = new URL(SG_BLOG_URL)
+
+    if (IDE === CodyIDE.VSCode) {
+        // Examples of version:
+        // 1.24.3 (stable), 1.25.123143 (pre-release)
+        const versionNums = version.split('.')
+        // NOTE: We do not generate blog post for pre-releases (odd minor number).
+        const minor = Number(versionNums[1]) % 2 === 0 ? versionNums[1] : `${Number(versionNums[1]) - 1}`
+        // Example: https://sourcegraph.com/blog/cody-vscode-1-24-0-release
+        blogURL.pathname += `cody-vscode-${versionNums[0]}-${minor}-0-release`
+    }
+
+    return blogURL.href
 }